+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 2023.03.01
+# 2024.08.11
+* Improved stability of td codes (filtering of noise from mean-fields).
+* Update for new API version of ELPA library.
+* Minor bug fixes and improvements.
+
+# 2024.03.01
 * Minor improvements of inittype=0 mode.
 * New API_VERSION=20240131. Users can write wave functions from td code to file via logger.h and define new wdata variables via logger.h
 * Minor improvements in the HIPMODE version.